Added Command
Added: "org.blockserver.command" package
Added: "CommandType" class with all the commands for Blockserver inside
of it.

Edited: "ConsoleListener" with the fist command added into it (stop)
although it does not function, it will register.

Edited: "Server" class to register the ConsoleListener so commands will
get registered

Edited "ServerConfig" class so it used the "motd" property so it showed
in the ping, dont know why I did this. Feel free to revert it.

Fixed: "run" class, the suppress warnings was being used incorrectly
(ALL) it is supposed to be "all".

TODO: Change CommandType to ENUM