Changed Commands
Added a "Command" Class with the function getCommand(String Command);
This can be used to retrieve the command from the console or player

Added a "CommandMessage" class so the default messages for a command can
be called swiftly and be changed quickly.

Edited the "ConsoleListener" Class so it is compatible with the new
Classes added.

Edited the "CommandType" class to a ENUM, that is why the Command class
exists so you can retrieve the CommandType from a String.

TODO: Add command arguments