Fixed loop glitch
Found out there was a glitch where nothing would happen after the
console listener was triggered because of the while(true) loop. Fixed it
by moving the ConsoleHandler to a thread inside of the Server.java
Class, also changed the if statement in the console listener to
(!line.isEmpty()) and put the code there, then put the return statement
in the else statement.